Large Scale Restructure
Adopted new file layout and transitioned my HTML-based Alpha pages into php. Adopted the structure from instructors examples, deleting and modifying code and items where necessary to fit the design of my site. More to come with the final setup. (This update was the culminations of edits that took place of the week and a half (11/17/21 - now).